agent/requirements.eval.txt
langchain-core
pydantic-ai[mcp]
langfuse
openai
nest_asyncio
aiohttp
python-dotenv
pandas
tqdm
pytest
openevals